+++ json_serializable/test/json_serializable_test.dart
@@ -29,6 +29,7 @@ const _expectedAnnotatedTests = {
   'BadOneNamed',
   'BadToFuncReturnType',
   'BadTwoRequiredPositional',
+  'CtorDefaultValueAndJsonKeyDefaultValue',
   'DefaultDoubleConstants',
   'DefaultWithConstObject',
   'DefaultWithDisallowNullRequiredClass',