+++ json_serializable/CHANGELOG.md
@@ -1,4 +1,8 @@
-## 4.1.1-dev
+## 5.0.0-dev
+
+- Use the default value for optional constructor parameters if
+  `JsonKey.defaultValue` is not provided. This could be a breaking behavior
+  change in generated code in some cases.
 
 ## 4.1.0